+++ src/l2-proxy-version.h
@@ -1,4 +1,4 @@
 #ifndef L2_PROXY_VERSION_H
 #define L2_PROXY_VERSION_H
-#define VERSION "1.0.4-c13817d"
+#define VERSION "1.0.4-6c4efa5"
 #endif